pyproject.toml
[tool.poetry]
name = "asbestos-dashboard-data"
version = "0.1.0"
description = "Python toolkit for preprocessing data for the Philadelphia City Controller's School District Asbestos Dashboard"
authors = ["Nick Hand <nick.hand@phila.gov>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^8.0.1"
loguru = "^0.5.3"
python-dotenv = "^0.19.0"
pandas = "^1.3.3"
openpyxl = "^3.0.9"
ipython = "^7.28.0"
jupyterlab = "^3.2.1"
jupyterlab-code-formatter = "^1.4.10"
geopandas = "^0.10.2"
boto3 = "^1.19.7"
schuylkill = "^0.1.1"
phl-courts-scraper = "^0.1.1"
ais2gpd = "^1.0.5"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9.3"

[tool.poetry.scripts]
asbestos-dashboard-data = "asbestos_dashboard_data.__main__:main"


